Remove todo suggesting adding an indirection in the TLR reference.
I really doubt that it would have been an improvement, because it is only
the slow path of BlazePool.claim that does ThreadLocal writes. For the fast
path, we only do a ThreadLocal read, and then the indirection would just
another pointer we would have to follow, before getting to the interesting
payload.