Fix a flaky test.
When the timeout given to claim is zero, then we should at most poll from the queue once.
However, if we at first got an object that then turned out to be invalid, the queue would be polled twice.
This is fixed by adding another clause to break out of the loop, if we got a slot that is invalid, and we have run out of time.