Claim newly allocated objects ahead of existing ones
An additional RefillPile has been added, which is used by the background allocation thread to publish newly allocated objects.
The slow-path of `Pool.claim` then attempts to pop from this RefillPile,
before consulting the live-queue.
The effect is that library clients recover faster from temporary allocation failures.
Objects that fail to allocate are still sent to the live-queue,
and thus end up "at the back" of the line for claims.
While successful allocations will be published through the newAllocations
RefillPile, and will thus be claimed ahead of existing live objects.
The RefillPile is non-blocking, so we still rely on the live-queue
for implementing the blocking behaviour of claim.

This fixes ##124.