More testing of unblocking behaviour
Since newly allocated objects may now be put into the stack of a RefillPile,
it is important to verify that blocked claim calls will become unblocked
by an allocation.
This turns out to be the case, because the slow path of claim is already
looping and only blocks for a short while.
This ensures that the newAllocations RefillPile is consulted fairly often.