Avoid hiding test failures in shutdown timeout
PoolTest.backgroundExpirationMustNotExpireObjectsThatAreClaimed has been failing.
However, the failure is manifested as a timeout when the test is shutting the pool down.
This is most likely because the claim object has not been released,
and this object not being released is most likely caused by the real test failure.
This fixes the shutdown timeout failure so we can see the real test failure.