Preparations for a threadless allocation process
The AllocatorProcess is now an AllocationController, and the
allocation process is now the factory that builds the controllers.
The pool builder will eventually take an AllocationProcess as a
configuration option.
The idea behind that is that such objects can eventually be shared
among pools, for an eventual shared background process feature.
But first a threadless mode needs to be implemented.
The tests are also rearranged to accommodate future diversification.