Introduce an _inline_ pool mode
The inline pooling mode does not start any background threads.
Work normally handled by the background thread is instead handled by
the threads that call into the pool, usually via `claim`.

There are currently 14 known test failures.
There is also a bunch of javadoc that needs to be written and updated.
This will be taken care of in the following commits.