Fix flaky ThreadedPoolTest.mustNotHoldOnToDeallocatedObjectsWhenLeakDetectionIsEnabled
When background expiration checking was enabled, it was possible for the test to race with the background expiration check.
This could cause slots to end up in the disregardPile or live queue, instead of the thread-local slot as intended.
The disregardPile or live queue would then maintain a strong reference to the object that we were hoping would be deallocated and all references purged, besides the weak reference in the test.
The fix is to disable background expiration checking in this test.