Implement an IdentityHashSet for the PreciseLeakDetector
The IdentityHashSet is based on open-addressing 2-choice hashing with 8-element buckets.
Conceptually, this is like bucketed cuckoo-hashing without the cuckoo-ing.
The 8-element buckets, and the addition of an 8-element stash for unresolvable conflicts, allow the table to have high utilisation.
The table array is a power-of-two size, so in practice utilisation will be between 45% to 90+%.
This means it'll take up less memory than the IdentityHashMap-as-a-Set we had before, in part due to not storing values, and in part to using a higher-utilisation hashing algorithm.
Theoretically, the 2-choice hashing scheme have higher access costs, due to computing two hashes and having at least two cache-misses, one from each sub-table, but this doesn't seem to be all that noticeable in the grand scheme of things.
The access performance is dramatically better than the 2-level chaining Bagwell trie we used before, by simply scaling to larger sizes.
The Bagwell trie, as implemented, didn't scale to more than 1000 elements, and relied entirely on chaining for capacities beyond that.