Add a setting to optimise for reduced memory usage
This is true by default, and avoids the object padding related to preventing false sharing.
This makes slot objects smaller by default.
However, when this setting is turned off, the objects are now padded to 128 bytes, instead of the previous 64 bytes.
The 128 matches the cache line size of many large-scale ARM systems, and is also sympathetic to the adjacent cache-line prefetch seen on x86_64 systems.