First draft of implementing a new blocking queue
This queue is implemented using a pair of modified unpadded MpmcUnboundedXaddArrayQueues from JCTools.

One queue is used for messages, while another is used for the wait-set.
Since the wait-set is a queue, we get fairness from our blocking methods.

A few modifications were necessary, in order to implement a wait-set using the queue.
Most importantly, the 'offer' method now returns the claimed producer index, and a pair of 'casEntry' and 'hasEntry' methods has been added, which allow blocking threads to manage their wait-set entries.

The queues have also gotten a 'count' method, which allow us to implement some inspection methods that Stormpot needs.