Fix data race in StackCompletion
When StackCompletion is used as a Flow.Publisher, new subscribers would be added to the stack before the onSubscribe method had been called.
This allowed onComplete callbacks to arrive before the subscriber had subscribed, and lead to NullPointerExceptions in the onComplete because the BaseSubscriber.subscription was null.

The fix is to call onSubscribe before adding the node to the stack.